move from 'except' to 'exception'
